requirements.txt
aiofiles==0.6.0
alembic==1.4.3
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
docker==4.4.1
h11==0.9.0
httpcore==0.11.1
httptools==0.1.1
httpx==0.15.4
idna==2.10
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.10.0
multidict==5.0.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pywin32==227
requests==2.25.1
rfc3986==1.4.0
sanic==20.9.1
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.3.22
urllib3==1.26.3
websocket-client==0.57.0
websockets==8.1
